backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0
alembic[tz]==1.13.1
opensearch-py==2.4.2
aiosmtplib==3.0.1
pytest==8.0.0
httpx==0.26.0 